Fix file uploads in edit form
I've removed explicit `request.form` passing to edit form in the form instance
creation. The reason for this is that if we pass `formdata` to the constructor
it considers that all the data it needs, whilst `request.form` contains no
information about uploaded files.

If we simply omit passing `request.form` around, Flask-WTF takes care about
filling the form with `request.form` as well as `request.files`, if there are
any.

Hence file uploads now work in the model edit.