Default to case-insensitive filtering and searching (MongoEngine).
The ability to do a case-insensitive search was added with
pull request #177. Preceding the search term with an
asterisk made the search case insensitive.

This commit makes case insensitive the default. In order to
do this while keeping case-sensitive capability, the
asterisk now means the opposite: case sensitive.

Making filtering and searching case insensitive by default
is meant to match the usual behavior of search boxes on
web sites and operating systems.