Make 'contains' the default string filter operator.
Upon adding a filter for a string field, 'equals' is the
default selected operator. The broader option, 'contains',
is more like the searches people typically do on web
site and operating system search boxes. Being broader,
contains is more generally useful for string fields than
equals.

This change places contains first in the list so it is
the default choice, preventing the user from having to
select it. The change applies to all three data engines
(MongoEngine, Peewee, and SQLAlchemy).