Fix custom field classes with Bootstrap3
Currently the custom class is created inside the call to `field()` so we
can't actually check in the template whether the field or widget has set
any custom classes.

With Bootstrap3, the `render_field` macro overwrites the `class` anyway,
so we have to define the widget's CSS class in the Python class level
and look it up before we actually create the object.