Add documentation and example for using flask-security-too with Flask-Admin.
Flask-Security-Too is a maintained and improved fork of the original Flask-Security.

The example is more full-functioned - providing all major views adapted to the current
Flask-Admin templates.

The example also makes use of the new more batteries included patterns introduced in recent
versions of Flask-Security-Too.

It also shows how to localize - utilizing a new jinja method to allow the new
templates to access existing flask-security-too translations.