Use pytest fixtures instead of setup functions
This allows us to have (almost) a single place where we have to think
about Babel (in the `babel`) fixture.

This should make it easier for us to, in the future, run tests without
babel or flask-babel installed, to validate flask-admin for users
without that package. This is important because the `flask_admin.babel`
module has two clearly different codepaths depending on whether the
package is installed.

Importantly, because of the fixture chaining (`admin` requires `babel`),
it also means that `babel` (if available) is always configured for every
flask app instance for every test.