Fix JS elements regression in modal windows
In #2407 we removed some seemingly-duplicate network requests for
`bs4_modal.js` every time a modal window is loaded.

This would have been fine if that file didn't not include any
side-effects, but it did: it re-applied 'global styling' every time the
JS file is loaded.

The goal here was to process 'styling' for any dynamica/interactive
elements, such as map widgets, and fields that use 'select2' or
'x-editable'.

When we removed these 'extra' JS calls, the modal windows stopped
getting these interactive elements enabled/setup.

This patch restores that functionality, while still retaining the goal
of PR #2407: to prevent duplicate network requests for a file.

We load `bs<x>_modal.js` on all pages that can display modals; this file
now only sets up the event listener that is responsible for making sure
a modal is cleared out/refreshed between multiple modal displays.

The logic for applying styling to the window is now added inline to all
of the relevant modal components.

A further improvement on what we had before is that now, the call to add
styling is scoped specifically to the modal. This prevents the issue
described in #2351, where map elements may show up multiple times on
the page because `applyGlobalStyles` is called multiple times.