Add tests-constraints.txt
appengine-python-standard is becoming slightly out-of-date; it requires
urllib3 < 2.0.0.

Our tests dependencies want to install a version above 2.0.0, and with
`constrain_package_deps` enabled this then causes tox to be unable to
install flask-admin's dependencies because they conflict with the test
dependencies.

Using a constraints file we can force tox to install urllib3<2 when
installing test deps, which then allows flask-admin to be installed with
no dependency conflicts.