Remove (flask-)mongoengine support
Support for `mongoengine` relies heavily on `flask-mongoengine` to
provide wtforms integration. Unfortunately, `flask-mongoengine` has
not seen much activity of support over the last two years, and is no
longer compatible with recent versions of our core dependencies (eg
flask).

Given this, it feels inappropriate and undesirable for us to try to
continue to support the functionality that project brings, as it will
increasingly apply a higher maintenance burden on us and eventually
likely cause compatibility issues for Flask-Admin users generally (if
someone wants to use both sqlalchemy and mongoengine views in
Flask-Admin, they'd be forced onto the lowest common denominator
dependency versions).