Don't (try to) hide the `reset filters` button
This lookup doesn't work because it's doing a full "class equals" check,
but the class is "btn btn-primary".

Also, because the page doesn't refresh when individual filters are
removed, if the "Reset filters" button is removed altogether then users
have no way to "apply" the removal of all filters. So we actually need
to keep this button around when the last filter is removed (it remains a
probably-weird user interaction, but I'm maintaining that for now - not
because it's good, but just to not change the existing behaviour)