Tweak testing instructions and fix devcontainer for uv
Recommend running tests using the devcontainer docker-compose, which
provides postgis+azurite+mongo as backing services automatically.

This also updates the devcontainer configuration to use `uv` to install
project dependencies. We provide a custom override path for the virtual
environment inside the devcontainer because, by default, the `.venv` of
the host will be mounted into the container. Editing this from within
the container can break the host's venv.